feat: add ES6 modules, start to implement caret position observer
