+++ src/index.js
@@ -1,304 +1,19 @@
 import './index.html';
 import './style.scss';
+import {
+  keyMapEn, keyMapRu, keyMapEnShift, keyMapRuShift, specialArr,
+} from './modules/alphabet.js';
+import makeRow from './modules/createRow.js';
 
-const body = document.querySelector('body');
-let language = 'en';
-body.innerHTML = `
+document.querySelector('body').innerHTML = `
                   <textarea class="textarea" name="textarea" id="textarea" cols="100" rows="20"></textarea>
                   <div id="keyboard"></div>
                   <h2 class="descr">Keyboard was made on Windows</h2>
                   <h2 class="descr">Click left Ctrl + left Alt to change language</h2>
                   `;
-const keyCode = ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8', 'Digit9', 'Digit0', 'Minus', 'Equal', 'Backspace', 'Tab', 'KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT', 'KeyY', 'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Backslash', 'Delete', 'CapsLock', 'KeyA', 'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Enter', 'ShiftLeft', 'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'Slash', 'ArrowUp', 'ShiftRight', 'ControlLeft', 'MetaLeft', 'AltLeft', 'Space', 'AltRight', 'ArrowLeft', 'ArrowDown', 'ArrowRight', 'ControlRight'];
 
-const keyMapEn = {
-  Backquote: '`',
-  Digit1: '1',
-  Digit2: '2',
-  Digit3: '3',
-  Digit4: '4',
-  Digit5: '5',
-  Digit6: '6',
-  Digit7: '7',
-  Digit8: '8',
-  Digit9: '9',
-  Digit0: '0',
-  Minus: '-',
-  Equal: '=',
-  Backspace: 'Backspace',
-  Tab: 'Tab',
-  KeyQ: 'q',
-  KeyW: 'w',
-  KeyE: 'e',
-  KeyR: 'r',
-  KeyT: 't',
-  KeyY: 'y',
-  KeyU: 'u',
-  KeyI: 'i',
-  KeyO: 'o',
-  KeyP: 'p',
-  BracketLeft: '[',
-  BracketRight: ']',
-  Backslash: '\\',
-  CapsLock: 'CapsLock',
-  KeyA: 'a',
-  KeyS: 's',
-  KeyD: 'd',
-  KeyF: 'f',
-  KeyG: 'g',
-  KeyH: 'h',
-  KeyJ: 'j',
-  KeyK: 'k',
-  KeyL: 'l',
-  Semicolon: ';',
-  Quote: "'",
-  Enter: 'Enter',
-  ShiftLeft: 'Shift',
-  KeyZ: 'z',
-  KeyX: 'x',
-  KeyC: 'c',
-  KeyV: 'v',
-  KeyB: 'b',
-  KeyN: 'n',
-  KeyM: 'm',
-  Comma: ',',
-  Period: '.',
-  Slash: '/',
-  ShiftRight: 'Shift',
-  ArrowUp: '⇧',
-  ControlLeft: 'Ctrl',
-  MetaLeft: 'Win',
-  AltLeft: 'Alt',
-  Space: ' ',
-  AltRight: 'Alt',
-  ControlRight: 'Ctrl',
-  ArrowLeft: '⇦',
-  ArrowDown: '⇩',
-  ArrowRight: '⇨',
-  Delete: 'del',
-};
-const specialArr = ['Backspace', 'Tab', 'CapsLock', 'Enter', 'ShiftLeft', 'ShiftRight', 'ControlLeft', 'MetaLeft', 'AltLeft', 'AltRight', 'ControlRight', 'Delete'];
-const keyMapEnShift = {
-  ShiftLeft: 'Shift',
-  Backquote: '~',
-  Digit1: '!',
-  Digit2: '@',
-  Digit3: '#',
-  Digit4: '$',
-  Digit5: '%',
-  Digit6: '^',
-  Digit7: '&',
-  Digit8: '*',
-  Digit9: '(',
-  Digit0: ')',
-  Minus: '_',
-  Equal: '+',
-  Backspace: 'Backspace',
-  Tab: 'Tab',
-  KeyQ: 'Q',
-  KeyW: 'W',
-  KeyE: 'E',
-  KeyR: 'R',
-  KeyT: 'T',
-  KeyY: 'Y',
-  KeyU: 'U',
-  KeyI: 'I',
-  KeyO: 'O',
-  KeyP: 'P',
-  BracketLeft: '{',
-  BracketRight: '}',
-  Backslash: '|',
-  CapsLock: 'CapsLock',
-  KeyA: 'A',
-  KeyS: 'S',
-  KeyD: 'D',
-  KeyF: 'F',
-  KeyG: 'G',
-  KeyH: 'H',
-  KeyJ: 'J',
-  KeyK: 'K',
-  KeyL: 'L',
-  Semicolon: ':',
-  Quote: '"',
-  Enter: 'Enter',
-  KeyZ: 'Z',
-  KeyX: 'X',
-  KeyC: 'C',
-  KeyV: 'V',
-  KeyB: 'B',
-  KeyN: 'N',
-  KeyM: 'M',
-  Comma: '<',
-  Period: '>',
-  Slash: '?',
-  ShiftRight: 'Shift',
-  ArrowUp: '⇧',
-  ControlLeft: 'Ctrl',
-  MetaLeft: 'Win',
-  AltLeft: 'Alt',
-  Space: ' ',
-  AltRight: 'Alt',
-  ControlRight: 'Ctrl',
-  ArrowLeft: '⇦',
-  ArrowDown: '⇩',
-  ArrowRight: '⇨',
-  Delete: 'del',
-};
-
-const keyMapRu = {
-  Backquote: 'ё',
-  Digit1: '1',
-  Digit2: '2',
-  Digit3: '3',
-  Digit4: '4',
-  Digit5: '5',
-  Digit6: '6',
-  Digit7: '7',
-  Digit8: '8',
-  Digit9: '9',
-  Digit0: '0',
-  Minus: '-',
-  Equal: '=',
-  Backspace: 'Backspace',
-  Tab: 'Tab',
-  KeyQ: 'й',
-  KeyW: 'ц',
-  KeyE: 'у',
-  KeyR: 'к',
-  KeyT: 'е',
-  KeyY: 'н',
-  KeyU: 'г',
-  KeyI: 'ш',
-  KeyO: 'щ',
-  KeyP: 'з',
-  BracketLeft: 'х',
-  BracketRight: 'ъ',
-  Backslash: '\\',
-  CapsLock: 'CapsLock',
-  KeyA: 'ф',
-  KeyS: 'ы',
-  KeyD: 'в',
-  KeyF: 'а',
-  KeyG: 'п',
-  KeyH: 'р',
-  KeyJ: 'о',
-  KeyK: 'л',
-  KeyL: 'д',
-  Semicolon: 'ж',
-  Quote: 'э',
-  Enter: 'Enter',
-  ShiftLeft: 'Shift',
-  KeyZ: 'я',
-  KeyX: 'ч',
-  KeyC: 'с',
-  KeyV: 'м',
-  KeyB: 'и',
-  KeyN: 'т',
-  KeyM: 'ь',
-  Comma: 'б',
-  Period: 'ю',
-  Slash: '.',
-  ShiftRight: 'Shift',
-  ArrowUp: '⇧',
-  ControlLeft: 'Ctrl',
-  MetaLeft: 'Win',
-  AltLeft: 'Alt',
-  Space: ' ',
-  AltRight: 'Alt',
-  ControlRight: 'Ctrl',
-  ArrowLeft: '⇦',
-  ArrowDown: '⇩',
-  ArrowRight: '⇨',
-  Delete: 'del',
-};
-const keyMapRuShift = {
-  ShiftLeft: 'Shift',
-  Backquote: 'Ё',
-  Digit1: '!',
-  Digit2: '"',
-  Digit3: '№',
-  Digit4: ';',
-  Digit5: '%',
-  Digit6: ':',
-  Digit7: '?',
-  Digit8: '*',
-  Digit9: '(',
-  Digit0: ')',
-  Minus: '_',
-  Equal: '+',
-  Backspace: 'Backspace',
-  Tab: 'Tab',
-  KeyQ: 'Й',
-  KeyW: 'Ц',
-  KeyE: 'У',
-  KeyR: 'К',
-  KeyT: 'Е',
-  KeyY: 'Н',
-  KeyU: 'Г',
-  KeyI: 'Ш',
-  KeyO: 'Щ',
-  KeyP: 'З',
-  BracketLeft: 'Х',
-  BracketRight: 'Ъ',
-  Backslash: '/',
-  CapsLock: 'CapsLock',
-  KeyA: 'Ф',
-  KeyS: 'Ы',
-  KeyD: 'В',
-  KeyF: 'А',
-  KeyG: 'П',
-  KeyH: 'Р',
-  KeyJ: 'О',
-  KeyK: 'Л',
-  KeyL: 'Д',
-  Semicolon: 'Ж',
-  Quote: 'Э',
-  Enter: 'Enter',
-  KeyZ: 'Я',
-  KeyX: 'Ч',
-  KeyC: 'С',
-  KeyV: 'М',
-  KeyB: 'И',
-  KeyN: 'Т',
-  KeyM: 'Ь',
-  Comma: 'Б',
-  Period: 'Ю',
-  Slash: ',',
-  ShiftRight: 'Shift',
-  ArrowUp: '⇧',
-  ControlLeft: 'Ctrl',
-  MetaLeft: 'Win',
-  AltLeft: 'Alt',
-  Space: ' ',
-  AltRight: 'Alt',
-  ControlRight: 'Ctrl',
-  ArrowLeft: '⇦',
-  ArrowDown: '⇩',
-  ArrowRight: '⇨',
-  Delete: 'del',
-};
-
-function makeRow(start, finish, object) {
-  let out = '';
-  for (let i = start; i <= finish; i += 1) {
-    if (keyCode[i] === 'Backspace' || keyCode[i] === 'ShiftLeft') {
-      out += `<div data-code="${keyCode[i]}" class="key backspace">${object[keyCode[i]]}</div>`;
-    } else if (keyCode[i] === 'Tab') {
-      out += `<div data-code="${keyCode[i]}" class="key small">${object[keyCode[i]]}</div>`;
-    } else if (keyCode[i] === 'ShiftRight') {
-      out += `<div data-code="${keyCode[i]}" class="key shiftright">${object[keyCode[i]]}</div>`;
-    } else if (keyCode[i] === 'CapsLock') {
-      out += `<div data-code="${keyCode[i]}" class="key medium">${object[keyCode[i]]}</div>`;
-    } else if (keyCode[i] === 'Enter') {
-      out += `<div data-code="${keyCode[i]}" class="key large">${object[keyCode[i]]}</div>`;
-    } else if (keyCode[i] === 'Space') {
-      out += `<div data-code="${keyCode[i]}" class="key extralarge">${object[keyCode[i]]}</div>`;
-    } else {
-      out += `<div data-code="${keyCode[i]}" class="key">${object[keyCode[i]]}</div>`;
-    }
-  }
-  document.querySelector('#keyboard').innerHTML += `<div class="row">${out}</div>`;
-}
+let cursorPosition = 0;
+let language = 'en';
 
 function init(obj) {
   document.querySelector('#keyboard').innerHTML = '';
@@ -307,15 +22,22 @@ function init(obj) {
   makeRow(29, 41, obj);
   makeRow(42, 54, obj);
   makeRow(55, 63, obj);
+  language = obj.Language;
 }
-
-init(keyMapEn);
+init(JSON.parse(localStorage.getItem('object')) || keyMapEn);
 
 const textArea = document.querySelector('.textarea');
 const keys = document.querySelectorAll('.key');
 const leftShift = document.querySelector('[data-code="ShiftLeft"]');
 const rightShift = document.querySelector('[data-code="ShiftRight"]');
 const caps = document.querySelector('[data-code="CapsLock"]');
+
+function setCaretPosition(ctrl, start, end) {
+  if (ctrl.setSelectionRange) {
+    ctrl.focus();
+    ctrl.setSelectionRange(start, end);
+  }
+}
 function changeKeys(obj) {
   for (let i = 0; i < keys.length; i += 1) {
     keys[i].innerText = obj[keys[i].dataset.code];
@@ -345,8 +67,10 @@ function keyUpHandler(event) {
     }
   }
 }
+setCaretPosition(textArea, textArea.value.length, textArea.value.length);
 
 function keyDownHandler(event) {
+  cursorPosition = textArea.selectionStart;
   event.preventDefault();
   if (event.key === 'Shift') {
     if (language === 'ru') {
@@ -363,13 +87,21 @@ function keyDownHandler(event) {
     }
   }
   if (event.code === 'Backspace') {
-    textArea.value = textArea.value.slice(0, textArea.value.length - 1);
+    if (cursorPosition > 0) {
+      textArea.value = textArea.value.slice(0, cursorPosition - 1)
+    + textArea.value.slice(cursorPosition);
+      setCaretPosition(textArea, cursorPosition - 1, cursorPosition - 1);
+      cursorPosition -= 1;
+    }
   }
   if (event.code === 'Enter') {
-    textArea.value += '\n';
+    textArea.value = `${textArea.value.slice(0, cursorPosition)}\n${
+      textArea.value.slice(cursorPosition)}`;
+    cursorPosition += 1;
+    setCaretPosition(textArea, cursorPosition, cursorPosition);
   }
   if (event.code === 'Tab') {
-    textArea.value += '    ';
+    textArea.value += '\t';
   }
   if (event.code === 'CapsLock') {
     caps.classList.toggle('active');
@@ -388,10 +120,14 @@ function keyDownHandler(event) {
     activeKey.classList.add('active');
     if (specialArr.indexOf(event.code) === -1) {
       if (event.getModifierState('CapsLock')) {
-        textArea.value += activeKey.textContent.toUpperCase();
+        textArea.value = `${textArea.value.slice(0, cursorPosition)}${activeKey.textContent.toUpperCase()}${
+          textArea.value.slice(cursorPosition)}`;
       } else {
-        textArea.value += activeKey.textContent;
+        textArea.value = `${textArea.value.slice(0, cursorPosition)}${activeKey.textContent}${
+          textArea.value.slice(cursorPosition)}`;
       }
+      cursorPosition += 1;
+      setCaretPosition(textArea, cursorPosition, cursorPosition);
     }
   }
 }
@@ -429,6 +165,8 @@ function changeShift(event) {
 function virtualKeysHandler() {
   keys.forEach((key) => {
     key.addEventListener('click', (event) => {
+      cursorPosition = textArea.selectionStart;
+      textArea.focus();
       keys.forEach((item) => {
         if (item.dataset.code !== 'CapsLock') {
           item.classList.remove('active');
@@ -438,13 +176,18 @@ function virtualKeysHandler() {
         textArea.value += event.target.textContent;
       } else {
         if (event.target.dataset.code === 'Tab') {
-          textArea.value += '    ';
+          textArea.value += '\t';
         }
         if (event.target.dataset.code === 'Enter') {
           textArea.value += '\n';
         }
         if (event.target.dataset.code === 'Backspace') {
-          textArea.value = textArea.value.slice(0, textArea.value.length - 1);
+          if (cursorPosition > 0) {
+            textArea.value = textArea.value.slice(0, cursorPosition - 1)
+          + textArea.value.slice(cursorPosition);
+            setCaretPosition(textArea, cursorPosition - 1, cursorPosition - 1);
+            cursorPosition -= 1;
+          }
         }
         if (event.target.dataset.code === 'CapsLock') {
           event.target.classList.toggle('active');
@@ -472,9 +215,20 @@ function changeLanguage(event) {
     if (language === 'ru') {
       changeKeys(keyMapEn);
       language = 'en';
+      const currentLang = JSON.stringify(keyMapEn);
+      localStorage.setItem('object', currentLang);
     } else {
       changeKeys(keyMapRu);
       language = 'ru';
+      const currentLang = JSON.stringify(keyMapRu);
+      localStorage.setItem('object', currentLang);
+    }
+    if (caps.classList.contains('active')) {
+      for (let i = 0; i < keys.length; i += 1) {
+        if (keys[i].innerText.length === 1) {
+          keys[i].innerText = keys[i].innerText.toUpperCase();
+        }
+      }
     }
   }
 }
