+++ src/modules/alphabet.js
@@ -1,0 +1,277 @@
+const keyCode = ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8', 'Digit9', 'Digit0', 'Minus', 'Equal', 'Backspace', 'Tab', 'KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT', 'KeyY', 'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Backslash', 'Delete', 'CapsLock', 'KeyA', 'KeyS', 'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Enter', 'ShiftLeft', 'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'Slash', 'ArrowUp', 'ShiftRight', 'ControlLeft', 'MetaLeft', 'AltLeft', 'Space', 'AltRight', 'ArrowLeft', 'ArrowDown', 'ArrowRight', 'ControlRight'];
+
+const keyMapEn = {
+  Language: 'en',
+  Backquote: '`',
+  Digit1: '1',
+  Digit2: '2',
+  Digit3: '3',
+  Digit4: '4',
+  Digit5: '5',
+  Digit6: '6',
+  Digit7: '7',
+  Digit8: '8',
+  Digit9: '9',
+  Digit0: '0',
+  Minus: '-',
+  Equal: '=',
+  Backspace: 'Backspace',
+  Tab: 'Tab',
+  KeyQ: 'q',
+  KeyW: 'w',
+  KeyE: 'e',
+  KeyR: 'r',
+  KeyT: 't',
+  KeyY: 'y',
+  KeyU: 'u',
+  KeyI: 'i',
+  KeyO: 'o',
+  KeyP: 'p',
+  BracketLeft: '[',
+  BracketRight: ']',
+  Backslash: '\\',
+  CapsLock: 'CapsLock',
+  KeyA: 'a',
+  KeyS: 's',
+  KeyD: 'd',
+  KeyF: 'f',
+  KeyG: 'g',
+  KeyH: 'h',
+  KeyJ: 'j',
+  KeyK: 'k',
+  KeyL: 'l',
+  Semicolon: ';',
+  Quote: "'",
+  Enter: 'Enter',
+  ShiftLeft: 'Shift',
+  KeyZ: 'z',
+  KeyX: 'x',
+  KeyC: 'c',
+  KeyV: 'v',
+  KeyB: 'b',
+  KeyN: 'n',
+  KeyM: 'm',
+  Comma: ',',
+  Period: '.',
+  Slash: '/',
+  ShiftRight: 'Shift',
+  ArrowUp: '⇧',
+  ControlLeft: 'Ctrl',
+  MetaLeft: 'Win',
+  AltLeft: 'Alt',
+  Space: ' ',
+  AltRight: 'Alt',
+  ControlRight: 'Ctrl',
+  ArrowLeft: '⇦',
+  ArrowDown: '⇩',
+  ArrowRight: '⇨',
+  Delete: 'del',
+};
+
+const specialArr = ['Backspace', 'Tab', 'CapsLock', 'Enter', 'ShiftLeft', 'ShiftRight', 'ControlLeft', 'MetaLeft', 'AltLeft', 'AltRight', 'ControlRight', 'Delete'];
+
+const keyMapEnShift = {
+  ShiftLeft: 'Shift',
+  Backquote: '~',
+  Digit1: '!',
+  Digit2: '@',
+  Digit3: '#',
+  Digit4: '$',
+  Digit5: '%',
+  Digit6: '^',
+  Digit7: '&',
+  Digit8: '*',
+  Digit9: '(',
+  Digit0: ')',
+  Minus: '_',
+  Equal: '+',
+  Backspace: 'Backspace',
+  Tab: 'Tab',
+  KeyQ: 'Q',
+  KeyW: 'W',
+  KeyE: 'E',
+  KeyR: 'R',
+  KeyT: 'T',
+  KeyY: 'Y',
+  KeyU: 'U',
+  KeyI: 'I',
+  KeyO: 'O',
+  KeyP: 'P',
+  BracketLeft: '{',
+  BracketRight: '}',
+  Backslash: '|',
+  CapsLock: 'CapsLock',
+  KeyA: 'A',
+  KeyS: 'S',
+  KeyD: 'D',
+  KeyF: 'F',
+  KeyG: 'G',
+  KeyH: 'H',
+  KeyJ: 'J',
+  KeyK: 'K',
+  KeyL: 'L',
+  Semicolon: ':',
+  Quote: '"',
+  Enter: 'Enter',
+  KeyZ: 'Z',
+  KeyX: 'X',
+  KeyC: 'C',
+  KeyV: 'V',
+  KeyB: 'B',
+  KeyN: 'N',
+  KeyM: 'M',
+  Comma: '<',
+  Period: '>',
+  Slash: '?',
+  ShiftRight: 'Shift',
+  ArrowUp: '⇧',
+  ControlLeft: 'Ctrl',
+  MetaLeft: 'Win',
+  AltLeft: 'Alt',
+  Space: ' ',
+  AltRight: 'Alt',
+  ControlRight: 'Ctrl',
+  ArrowLeft: '⇦',
+  ArrowDown: '⇩',
+  ArrowRight: '⇨',
+  Delete: 'del',
+};
+
+const keyMapRu = {
+  Language: 'ru',
+  Backquote: 'ё',
+  Digit1: '1',
+  Digit2: '2',
+  Digit3: '3',
+  Digit4: '4',
+  Digit5: '5',
+  Digit6: '6',
+  Digit7: '7',
+  Digit8: '8',
+  Digit9: '9',
+  Digit0: '0',
+  Minus: '-',
+  Equal: '=',
+  Backspace: 'Backspace',
+  Tab: 'Tab',
+  KeyQ: 'й',
+  KeyW: 'ц',
+  KeyE: 'у',
+  KeyR: 'к',
+  KeyT: 'е',
+  KeyY: 'н',
+  KeyU: 'г',
+  KeyI: 'ш',
+  KeyO: 'щ',
+  KeyP: 'з',
+  BracketLeft: 'х',
+  BracketRight: 'ъ',
+  Backslash: '\\',
+  CapsLock: 'CapsLock',
+  KeyA: 'ф',
+  KeyS: 'ы',
+  KeyD: 'в',
+  KeyF: 'а',
+  KeyG: 'п',
+  KeyH: 'р',
+  KeyJ: 'о',
+  KeyK: 'л',
+  KeyL: 'д',
+  Semicolon: 'ж',
+  Quote: 'э',
+  Enter: 'Enter',
+  ShiftLeft: 'Shift',
+  KeyZ: 'я',
+  KeyX: 'ч',
+  KeyC: 'с',
+  KeyV: 'м',
+  KeyB: 'и',
+  KeyN: 'т',
+  KeyM: 'ь',
+  Comma: 'б',
+  Period: 'ю',
+  Slash: '.',
+  ShiftRight: 'Shift',
+  ArrowUp: '⇧',
+  ControlLeft: 'Ctrl',
+  MetaLeft: 'Win',
+  AltLeft: 'Alt',
+  Space: ' ',
+  AltRight: 'Alt',
+  ControlRight: 'Ctrl',
+  ArrowLeft: '⇦',
+  ArrowDown: '⇩',
+  ArrowRight: '⇨',
+  Delete: 'del',
+};
+
+const keyMapRuShift = {
+  ShiftLeft: 'Shift',
+  Backquote: 'Ё',
+  Digit1: '!',
+  Digit2: '"',
+  Digit3: '№',
+  Digit4: ';',
+  Digit5: '%',
+  Digit6: ':',
+  Digit7: '?',
+  Digit8: '*',
+  Digit9: '(',
+  Digit0: ')',
+  Minus: '_',
+  Equal: '+',
+  Backspace: 'Backspace',
+  Tab: 'Tab',
+  KeyQ: 'Й',
+  KeyW: 'Ц',
+  KeyE: 'У',
+  KeyR: 'К',
+  KeyT: 'Е',
+  KeyY: 'Н',
+  KeyU: 'Г',
+  KeyI: 'Ш',
+  KeyO: 'Щ',
+  KeyP: 'З',
+  BracketLeft: 'Х',
+  BracketRight: 'Ъ',
+  Backslash: '/',
+  CapsLock: 'CapsLock',
+  KeyA: 'Ф',
+  KeyS: 'Ы',
+  KeyD: 'В',
+  KeyF: 'А',
+  KeyG: 'П',
+  KeyH: 'Р',
+  KeyJ: 'О',
+  KeyK: 'Л',
+  KeyL: 'Д',
+  Semicolon: 'Ж',
+  Quote: 'Э',
+  Enter: 'Enter',
+  KeyZ: 'Я',
+  KeyX: 'Ч',
+  KeyC: 'С',
+  KeyV: 'М',
+  KeyB: 'И',
+  KeyN: 'Т',
+  KeyM: 'Ь',
+  Comma: 'Б',
+  Period: 'Ю',
+  Slash: ',',
+  ShiftRight: 'Shift',
+  ArrowUp: '⇧',
+  ControlLeft: 'Ctrl',
+  MetaLeft: 'Win',
+  AltLeft: 'Alt',
+  Space: ' ',
+  AltRight: 'Alt',
+  ControlRight: 'Ctrl',
+  ArrowLeft: '⇦',
+  ArrowDown: '⇩',
+  ArrowRight: '⇨',
+  Delete: 'del',
+};
+
+export {
+  keyCode, keyMapEn, keyMapRu, keyMapEnShift, keyMapRuShift, specialArr,
+};
